Renamed the syslog package to logcollector
